refactor out call page
