+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CallPage/CallPage.tsx
@@ -10,29 +10,26 @@ import {
   useGridApiRef,
 } from '@mui/x-data-grid-pro';
 import _ from 'lodash';
-import React, {FC, useCallback, useEffect, useMemo} from 'react';
+import React, {useCallback, useEffect, useMemo} from 'react';
 import {useHistory} from 'react-router-dom';
 
-import {parseRef} from '../../../../../react';
-import {Timestamp} from '../../../../Timestamp';
-import {Call} from '../../Browse2/callTree';
-import {parseRefMaybe, SmallRef} from '../../Browse2/SmallRef';
-import {useWeaveflowCurrentRouteContext} from '../context';
-import {CallStatusCodeChip} from './common/CallStatusCodeChip';
-import {KeyValueTable} from './common/KeyValueTable';
-import {CenteredAnimatedLoader} from './common/Loader';
-import {OpVersionCategoryChip} from './common/OpVersionCategoryChip';
+import {parseRef} from '../../../../../../react';
+import {Call} from '../../../Browse2/callTree';
+import {SmallRef} from '../../../Browse2/SmallRef';
+import {useWeaveflowCurrentRouteContext} from '../../context';
+import {CenteredAnimatedLoader} from '../common/Loader';
+import {OpVersionCategoryChip} from '../common/OpVersionCategoryChip';
 import {
-  SimpleKeyValueTable,
   SimplePageLayout,
   SimplePageLayoutContext,
   SimplePageLayoutWithHeader,
-} from './common/SimplePageLayout';
-import {UnderConstruction} from './common/UnderConstruction';
-import {GroupedCalls} from './ObjectVersionPage';
-import {truncateID} from './util';
-import {useWeaveflowORMContext} from './wfInterface/context';
-import {WFCall} from './wfInterface/types';
+} from '../common/SimplePageLayout';
+import {UnderConstruction} from '../common/UnderConstruction';
+import {truncateID} from '../util';
+import {useWeaveflowORMContext} from '../wfInterface/context';
+import {WFCall} from '../wfInterface/types';
+import {CallDetails} from './CallDetails';
+import {CallOverview} from './CallOverview';
 
 // % of screen to give the trace view in horizontal mode
 const TRACE_PCT = 40;
@@ -695,138 +692,4 @@ const useCallFlattenedTraceTree = (call: WFCall) => {
 
     return {rows, expandKeys};
   }, [call]);
-};
-
-const CallDetails: FC<{
-  wfCall: WFCall;
-}> = ({wfCall}) => {
-  const call = wfCall.rawCallSpan();
-  const inputKeys =
-    call.inputs._keys ??
-    Object.entries(call.inputs)
-      .filter(([k, c]) => c != null && !k.startsWith('_'))
-      .map(([k, c]) => k);
-  const inputs = _.fromPairs(inputKeys.map(k => [k, call.inputs[k]]));
-
-  const callOutput = call.output ?? {};
-  const outputKeys =
-    callOutput._keys ??
-    Object.entries(callOutput)
-      .filter(([k, c]) => c != null && (k === '_result' || !k.startsWith('_')))
-      .map(([k, c]) => k);
-  const output = _.fromPairs(outputKeys.map(k => [k, callOutput[k]]));
-
-  return (
-    <Box
-      style={{
-        width: '100%',
-        height: '100%',
-        overflowX: 'hidden',
-        overflowY: 'auto',
-        // padding: 2,
-      }}>
-      <Box
-        style={{
-          width: '100%',
-          display: 'flex',
-          flexDirection: 'column',
-          height: '100%',
-          gap: 1,
-          paddingTop: '8px',
-        }}>
-        {Object.keys(inputs).length > 0 && (
-          <Box
-            sx={{
-              flex: '0 0 auto',
-              p: 2,
-            }}>
-            <KeyValueTable
-              headerTitle="Inputs"
-              data={
-                // TODO: Consider bringing back openai chat input/output
-                inputs
-              }
-            />
-          </Box>
-        )}
-        {Object.keys(output).length > 0 && (
-          <Box
-            sx={{
-              flex: '0 0 auto',
-              p: 2,
-            }}>
-            <KeyValueTable
-              headerTitle="Output"
-              data={
-                // TODO: Consider bringing back openai chat input/output
-                output
-              }
-            />
-          </Box>
-        )}
-      </Box>
-    </Box>
-  );
-};
-
-const CallOverview: React.FC<{
-  wfCall: WFCall;
-}> = ({wfCall}) => {
-  const call = wfCall.rawCallSpan();
-  const opCategory = wfCall.opVersion()?.opCategory();
-  const childCalls = wfCall.childCalls().filter(c => {
-    return c.opVersion() != null;
-  });
-  const attributes = _.fromPairs(
-    Object.entries(call.attributes ?? {}).filter(
-      ([k, a]) => !k.startsWith('_') && a != null
-    )
-  );
-  const summary = _.fromPairs(
-    Object.entries(call.summary ?? {}).filter(
-      ([k, a]) => !k.startsWith('_') && k !== 'latency_s' && a != null
-    )
-  );
-
-  return (
-    <SimpleKeyValueTable
-      data={{
-        Operation:
-          parseRefMaybe(call.name) != null ? (
-            <SmallRef objRef={parseRefMaybe(call.name)!} wfTable="OpVersion" />
-          ) : (
-            call.name
-          ),
-        ...(opCategory
-          ? {
-              Category: <OpVersionCategoryChip opCategory={opCategory} />,
-            }
-          : {}),
-        Status: (
-          <CallStatusCodeChip statusCode={call.status_code as any} showLabel />
-        ),
-        Called: <Timestamp value={call.timestamp / 1000} format="relative" />,
-        ...(call.summary.latency_s != null
-          ? {
-              Latency: call.summary.latency_s.toFixed(3) + 's',
-            }
-          : {}),
-        ...(call.exception ? {Exception: call.exception} : {}),
-        ...(childCalls.length > 0
-          ? {
-              'Child Calls': (
-                <GroupedCalls
-                  calls={childCalls}
-                  partialFilter={{
-                    parentId: wfCall.callID(),
-                  }}
-                />
-              ),
-            }
-          : {}),
-        ...(Object.keys(attributes).length > 0 ? {Attributes: attributes} : {}),
-        ...(Object.keys(summary).length > 0 ? {Summary: summary} : {}),
-      }}
-    />
-  );
 };
